Refactor: PD-126 JobPosting, ClosingSoon 컴포넌트 분리
